Update README.md
Hello,

Here is how I have started.  This would just be starting with the Project 8, I still think this needs to have the window size set.  When testing this my second image is taking a while to pop up....I think that is probably due to its size.   Let me know what happens when you look at it.

-Diana