Added the grayscale conversion
I saw the project asks to convert the image into grayscale then into sepia. I added the grayscale function and added it to the main function. I was testing the code chunk from the book for the sepia conversion and it wasn't doing anything, so I added the .setPixel() method into the if/else statements. I removed a few things like the "close the picture window to continue" just for now because I was testing the end result over and over and it was quicker at the time. I wanted to make sure you think this is working correctly too before we put it all together.