Update Ch7 P8P9
Looks good, I updated all the variables to match and added the section to lighten an image which appears to be working from my testing.  I also added the projects together into one file.

I am working on the darkening now.  Would you be able to start on the color filter function?
-Diana